refactor: 히스토리 로직 분리
