+++ src/main/java/roomescape/reservation/service/ReservationApplicationService.java
@@ -8,6 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.domain.AuthInfo;
 import roomescape.member.controller.AdminMemberReservationRequest;
 import roomescape.member.domain.Member;
+import roomescape.payment.domain.Payment;
 import roomescape.payment.service.PaymentService;
 import roomescape.payment.service.dto.PaymentRequest;
 import roomescape.reservation.controller.dto.ReservationPaymentRequest;
@@ -63,10 +64,10 @@ public class ReservationApplicationService {
 
         memberReservation.approve();
 
-        paymentService.pay(
+        Payment payment = paymentService.pay(
                 new PaymentRequest(reservationPaymentRequest.amount(), reservationPaymentRequest.orderId(),
                         reservationPaymentRequest.paymentKey()), memberReservation);
-
+        paymentService.createHistory(memberReservation, payment);
         return ReservationResponse.from(memberReservation);
     }
 
@@ -74,9 +75,10 @@ public class ReservationApplicationService {
     public ReservationResponse createMemberReservation(MemberReservationCreate memberReservationCreate) {
         MemberReservation memberReservation = reservationCommonService.create(
                 memberReservationCreate.toReservationCreate());
-        paymentService.pay(
+        Payment payment = paymentService.pay(
                 new PaymentRequest(memberReservationCreate.amount(), memberReservationCreate.orderId(),
                         memberReservationCreate.paymentKey()), memberReservation);
+        paymentService.createHistory(memberReservation, payment);
         return ReservationResponse.from(memberReservation);
     }
