+++ src/test/java/roomescape/payment/service/PaymentServiceTest.java
@@ -126,32 +126,6 @@ class PaymentServiceTest extends ServiceTest {
         assertThat(payment.isPresent()).isFalse();
     }
 
-    @DisplayName("결제에 성공하면, 히스토리가 생성된다.")
-    @Test
-    void createHistory() {
-        //given
-        String paymentKey = "tgen_20240528172021mxEG4";
-        String paymentType = "카드";
-        BigDecimal totalAmount = BigDecimal.valueOf(1000L);
-        PaymentRequest paymentRequest = new PaymentRequest(totalAmount, "MC45NTg4ODYxMzA5MTAz", paymentKey);
-        PaymentResponse okResponse =
-                new PaymentResponse(paymentKey, "DONE", "MC4wOTA5NzEwMjg3MjQ2", totalAmount, paymentType);
-        doReturn(okResponse).when(paymentClient).confirm(any());
-
-        //when
-        paymentService.pay(paymentRequest, memberReservation);
-
-        //then
-        Optional<PaymentHistory> paymentHistory = paymentHistoryRepository.findPaymentHistoryByPaymentKey(
-                paymentKey);
-
-        assertAll(
-                () -> assertThat(paymentHistory).isNotNull(),
-                () -> assertThat(paymentHistory.get().getPaymentType()).isEqualTo(PaymentType.from(paymentType)),
-                () -> assertThat(paymentHistory.get().getPaymentStatus()).isEqualTo(PaymentStatus.PAID)
-        );
-    }
-
     @DisplayName("결제 환불 시, 히스토리가 갱신된다.")
     @Test
     void updateHistory() {
